feat(analytics): share PostHog config and improve server delivery
Extract shared PostHog config/types into a single module, add server-side alias + flush helpers, and flush analytics before redirects/response end to reduce dropped events.
